+++ src/busca/dijkstra.py
@@ -1,8 +1,11 @@
 """Implementação do algoritmo de Dijkstra para o menor caminho em grafos."""
 
 from heapq import heappush, heappop
-from ..util import get_neighbors
 
+
+def get_neighbors(graph, node):
+    """Return neighbors from a node."""
+    return graph.get(node, [])
 
 def dijkstra(graph, start: int, goal: int) -> (int, float, [int]):
     """Busca em graph, um caminho entre start e goal usando Dijkstra."""
